renamed packages, created migration file
