--- models/user.go
@@ -1,7 +1,0 @@
-package models
-
-type User struct {
-	Id        string `json:"id"`
-	FirstName string `json:"firstName"`
-	LastName  string `json:"lastName"`
-}
