+++ models/users.go
@@ -1,0 +1,7 @@
+package models
+
+type User struct {
+	Id        string `json:"id"`
+	FirstName string `json:"firstName"`
+	LastName  string `json:"lastName"`
+}
