+++ router/router.go
@@ -2,7 +2,7 @@ package router
 
 import (
 	"github.com/gofiber/fiber/v2"
-	"github.com/rogerok/wflow-backend/api"
+	"github.com/rogerok/wflow-backend/handlers"
 
 	"github.com/gofiber/fiber/v2/middleware/logger"
 )
@@ -11,6 +11,6 @@ func SetupRouter(app *fiber.App) {
 	app.Use(logger.New())
 
 	apiv1 := app.Group("/api")
-	apiv1.Get("/user", api.HandleGetUsers)
-	apiv1.Get("/user/:id", api.HandleGetUser)
+	apiv1.Get("/user", handlers.HandleGetUsers)
+	apiv1.Get("/user/:id", handlers.HandleGetUser)
 }
